cloud_functions/transcript_to_report/requirements.txt
functions-framework==3.*
anthropic==0.40.*
google-api-python-client==2.*
google-auth==2.*
requests==2.*
